improved health check

+++ plotter/tensorboard_logger.py
@@ -8,11 +8,16 @@ from torch.utils.tensorboard import SummaryWriter
 class TensorBoard(SummaryWriter):
     def __init__(self, path: str):
         super(TensorBoard, self).__init__(path)
+        self.process = None
         try:
-            self.process = subprocess.Popen(["tensorboard", "--logdir="+path], shell=False)
-            threading.Thread(target=lambda: time.sleep(3) or webbrowser.open("http://localhost:6006")).start()
+            threading.Thread(target=lambda: time.sleep(3) or self.launch(path)).start()
         except FileNotFoundError:
             print("tensorboard is not installed")
 
+    def launch(self, path):
+        self.process = subprocess.Popen(["tensorboard", "--logdir=" + path], shell=False)
+        threading.Thread(target=lambda: time.sleep(1) or webbrowser.open("http://localhost:6006")).start()
+
     def __del__(self):
-        self.process.kill()
+        if self.process:
+            self.process.kill()
